Add base implementation of transforms and change base implementation of linear and nonlinear transforms
